+++ src/api-webhook.js
@@ -6,12 +6,11 @@ import log from "./log";
  * Handles web-hook callbacks from Github.
  *
  * @param {Object} settings:
- *                  - apps:       Collection of apps to start.
- *                  - manifest:   A manifest object.
+ *                  - mainApi:    The main API.
  *
  */
 export default (settings = {}) => {
-  const { apps } = settings;
+  const { mainApi } = settings;
   return {
     post(req, res) {
       // Extract data.
@@ -21,7 +20,7 @@ export default (settings = {}) => {
 
       // Match apps that reside within the repo that Github posted.
       const isRepoMatch = (app) => app.repo.name === repo && app.branch === branch;
-      const matchingApps = R.filter(isRepoMatch, apps);
+      const matchingApps = R.filter(isRepoMatch, mainApi.apps);
 
       // Update any matching apps.
       if (matchingApps.length > 0) {