+++ test/route.test.js
@@ -17,22 +17,29 @@ describe("route", function() {
     expect(route.parse("*/").domain).to.equal("*");
     expect(route.parse(" * ").domain).to.equal("*");
     expect(route.parse(" *  /").domain).to.equal("*");
+    expect(route.parse(" *  /path").domain).to.equal("*");
   });
 
 
-  it("has no path", () => {
-    expect(route.parse("*").path).to.equal(undefined);
-    expect(route.parse("*/ ").path).to.equal(undefined);
+  it("has a wildcard path (*)", () => {
+    expect(route.parse("*").path).to.equal("*");
+    expect(route.parse("*/").path).to.equal("*");
+    expect(route.parse("*/  ").path).to.equal("*");
+    expect(route.parse("*/*").path).to.equal("*");
+    expect(route.parse("*/*/").path).to.equal("*");
+    expect(route.parse("*/  *  ").path).to.equal("*");
+    expect(route.parse("domain.com/*").path).to.equal("*");
+    expect(route.parse("domain.com").path).to.equal("*");
   });
 
 
   describe("match", function() {
-    it.only("matches on specific domain (no path)", () => {
+    it("matches on specific domain (no path)", () => {
       expect(route.parse("www.foo.com").match("www.foo.com")).to.equal(true);
       expect(route.parse("www.foo.com").match("www.foo.com", "")).to.equal(true);
-
-      // expect(route.parse("www.foo.com/*").match("www.foo.com", "bar")).to.equal(true);
-      // expect(route.parse("www.foo.com").match("www.foo.com", "/bar")).to.equal(true);
+      expect(route.parse("www.foo.com/*").match("www.foo.com", "bar")).to.equal(true);
+      expect(route.parse("www.foo.com/*").match("www.foo.com", "/bar")).to.equal(true);
+      expect(route.parse("www.foo.com").match("www.foo.com", "/bar")).to.equal(true);
     });
 
     it("matches on specific domain (with path)", () => {
@@ -44,7 +51,6 @@ describe("route", function() {
     it("does not match on specific domain (no path)", () => {
       expect(route.parse("www.foo.com").match("www.bar.org")).to.equal(false);
       expect(route.parse("www.foo.com").match("www.bar.org", "foo")).to.equal(false);
-      expect(route.parse("www.foo.com").match("www.foo.com", "bar")).to.equal(false);
     });
 
     it("does not match on specific domain (with path)", () => {
@@ -62,12 +68,12 @@ describe("route", function() {
     it("matches on wildcard domain (*) with no path", () => {
       expect(route.parse("*").match("foo.com")).to.equal(true);
       expect(route.parse("*").match("bar.com")).to.equal(true);
+      expect(route.parse("*").match("bar.com", "path")).to.equal(true);
     });
 
     it("does not match on wildcard domain (*) with path", () => {
       expect(route.parse("*/foo").match("bar.com")).to.equal(false);
       expect(route.parse("*/foo").match("bar.com", "bar")).to.equal(false);
-      expect(route.parse("*").match("bar.com", "bar")).to.equal(false);
     });
   });
 });